Fix AzureDevOpsTimeline test failures: extract processor from SF host (#6492)
