+++ src/Telemetry/AzureDevOpsTimeline/Program.cs
@@ -64,6 +64,7 @@ internal static class Program
                         services.AddSingleton<ITimelineTelemetryRepository, KustoTimelineTelemetryRepository>();
                         services.AddSingleton<IBuildLogScraper, BuildLogScraper>();
                         services.AddSingleton<ExponentialRetry>();
+                        services.AddScoped<AzureDevOpsTimelineProcessor>();
                     });                    
             });
     }
