+++ src/Telemetry/Microsoft.DotNet.AzureDevOpsTimelineTests/AzureDevOpsTimelineTests.cs
@@ -38,10 +38,10 @@ public partial class AzureDevOpsTimelineTests
             });
         }
 
-        public static Func<IServiceProvider,AzureDevOpsTimeline> Controller(IServiceCollection collection)
+        public static Func<IServiceProvider,AzureDevOpsTimelineProcessor> Controller(IServiceCollection collection)
         {
-            collection.AddScoped<AzureDevOpsTimeline>();
-            return s => s.GetRequiredService<AzureDevOpsTimeline>();
+            collection.AddScoped<AzureDevOpsTimelineProcessor>();
+            return s => s.GetRequiredService<AzureDevOpsTimelineProcessor>();
         }
 
         [ConfigureAllParameters]
